move the vehicle information page
